Componentize LogDestinationIndicator and QueryFrequencyIndicator
