+++ frontend/components/QueryFrequencyIndicator/_styles.scss
@@ -1,0 +1,13 @@
+.query-frequency-indicator {
+  width: 100px;
+  display: flex;
+  padding: 8px 12px;
+
+  .icon {
+    padding-right: $pad-small;
+  }
+}
+
+.grey {
+  color: $ui-fleet-black-33;
+}
