+++ frontend/pages/queries/ManageQueriesPage/components/ManageAutomationsModal/ManageAutomationsModal.tsx
@@ -6,8 +6,8 @@ import Button from "components/buttons/Button";
 import InfoBanner from "components/InfoBanner/InfoBanner";
 import CustomLink from "components/CustomLink/CustomLink";
 import Checkbox from "components/forms/fields/Checkbox/Checkbox";
-import TooltipWrapper from "components/TooltipWrapper/TooltipWrapper";
-import Icon from "components/Icon/Icon";
+import QueryFrequencyIndicator from "components/QueryFrequencyIndicator/QueryFrequencyIndicator";
+import LogDestinationIndicator from "components/LogDestinationIndicator/LogDestinationIndicator";
 
 import { ISchedulableQuery } from "interfaces/schedulable_query";
 
@@ -70,7 +70,6 @@ const ManageAutomationsModal = ({
   // TODO: Error handling, if any
   const [errors, setErrors] = useState<{ [key: string]: string }>({});
 
-  console.log("MANAGEAUTOMATIONSMODAL automatedQueryIds", automatedQueryIds);
   const { queryItems, updateQueryItems } = useCheckboxListStateManagement(
     availableQueries || [],
     automatedQueryIds || []
@@ -79,11 +78,9 @@ const ManageAutomationsModal = ({
   const onSubmit = (evt: React.MouseEvent<HTMLFormElement> | KeyboardEvent) => {
     evt.preventDefault();
 
-    console.log("automatedQueryIds", automatedQueryIds);
     const newQueryIds: number[] = [];
     queryItems?.forEach((p) => p.isChecked && newQueryIds.push(p.id));
 
-    console.log("newQueryIds", newQueryIds);
     handleSubmit(newQueryIds);
   };
 
@@ -99,117 +96,6 @@ const ManageAutomationsModal = ({
       document.removeEventListener("keydown", listener);
     };
   });
-
-  const renderLogDestination = () => {
-    const readableLogDestination = () => {
-      switch (logDestination) {
-        case "filesystem":
-          return "Filesystem";
-        case "firehose":
-          return "Amazon Kinesis Data Firehose";
-        case "kinesis":
-          return "Amazon Kinesis Data Streams";
-        case "lambda":
-          return "AWS Lambda";
-        case "pubsub":
-          return "Google Cloud Pub/Sub";
-        case "kafta":
-          return "Apache Kafka";
-        case "stdout":
-          return "Standard output (stdout)";
-        case "":
-          return "Not configured";
-        default:
-          return logDestination;
-      }
-    };
-
-    const tooltipText = () => {
-      switch (logDestination) {
-        case "filesystem":
-          return `Each time a query runs, the data is sent to <br />
-            /var/log/osquery/osqueryd.snapshots.log <br />
-            in each host&apos;s filesystem.`;
-        case "firehose":
-          return `Each time a query runs, the data is sent to <br />
-            Amazon Kinesis Data Firehose.`;
-        case "kinesis":
-          return `Each time a query runs, the data is sent to <br />
-            Amazon Kinesis Data Streams.`;
-        case "lambda":
-          return `
-            Each time a query runs, the data <br />is sent to AWS Lambda.
-          `;
-        case "pubsub":
-          return `Each time a query runs, the data is <br />sent to Google Cloud Pub/Sub.`;
-        case "kafta":
-          return `Each time a query runs, the data <br />is sent to Apache Kafka.`;
-        case "stdout":
-          return `Each time a query runs, the data is sent to <br />
-            standard output (stdout) on the Fleet server.`;
-        case "":
-          return "Please configure a log destination.";
-        default:
-          return "No additional information is available about this log destination.";
-      }
-    };
-
-    return (
-      <TooltipWrapper tipContent={tooltipText()}>
-        {readableLogDestination()}
-      </TooltipWrapper>
-    );
-  };
-
-  const renderScheduleIndicator = ({
-    frequency,
-    checked,
-  }: IFrequencyIndicator) => {
-    const readableQueryFrequency = () => {
-      switch (frequency) {
-        case 0:
-          return "Never";
-        case 300:
-        case 600:
-        case 900:
-        case 1800: // 5, 10, 15, 30 minutes
-          return `${(frequency / 60).toString()} minutes`;
-        case 3600:
-          return "Hourly";
-        case 21600:
-        case 43200: // 6, 12 hours
-          return `${(frequency / 3600).toString()} hours`;
-        case 86400:
-          return "Daily";
-        case 604800:
-          return "Weekly";
-        default:
-          return "Unknown";
-      }
-    };
-
-    const frequencyIcon = () => {
-      switch (frequency) {
-        case 0:
-          return checked ? (
-            <Icon size="small" name="warning" />
-          ) : (
-            <Icon size="small" name="clock" color="ui-fleet-black-33" />
-          );
-        default:
-          return <Icon size="small" name="clock" />;
-      }
-    };
-    return (
-      <div
-        className={`${baseClass}__schedule-indicator
-        ${frequency === 0 && !checked && "grey"}`}
-      >
-        {frequencyIcon()}
-        {readableQueryFrequency()}
-      </div>
-    );
-  };
 
   return (
     <Modal
@@ -245,10 +131,10 @@ const ManageAutomationsModal = ({
                       >
                         {name}
                       </Checkbox>
-                      {renderScheduleIndicator({
-                        frequency: interval,
-                        checked: isChecked,
-                      })}
+                      <QueryFrequencyIndicator
+                        frequency={interval}
+                        checked={isChecked}
+                      />
                     </div>
                   );
                 })}
@@ -265,7 +151,7 @@ const ManageAutomationsModal = ({
             <strong>Log destination:</strong>
           </p>
           <div className={`${baseClass}__selection`}>
-            {renderLogDestination()}
+            <LogDestinationIndicator logDestination={logDestination} />
           </div>
           <div className={`${baseClass}__configure`}>
             Users with the admin role can&nbsp;
